+++ packages/terra-framework-docs/src/terra-dev-site/test/menu/menu/DefaultMenu.test.jsx
@@ -1,5 +1,7 @@
 import React from 'react';
 import Menu from 'terra-menu';
+import Button from 'terra-button';
+import IconCaretDown from 'terra-icon/lib/icon/IconCaretDown';
 
 class DefaultMenu extends React.Component {
   constructor(props) {
@@ -41,9 +43,7 @@ class DefaultMenu extends React.Component {
         >
           <Menu.Item text="Default Menu" id="TestContent" />
         </Menu>
-        <button type="button" id="default-button" onClick={this.handleButtonClick} ref={this.setButtonNode}>
-          Default Menu
-        </button>
+        <Button id="default-button" type="button" onClick={this.handleButtonClick} text="Default Menu" aria-haspopup icon={<IconCaretDown />} isReversed refCallback={this.setButtonNode} />
       </div>
     );
   }