+++ packages/terra-framework-docs/src/terra-dev-site/test/menu/menu/NonSelectableMenu.test.jsx
@@ -1,5 +1,7 @@
 import React from 'react';
 import Menu from 'terra-menu';
+import Button from 'terra-button';
+import IconCaretDown from 'terra-icon/lib/icon/IconCaretDown';
 
 class NonSelectableMenu extends React.Component {
   constructor(props) {
@@ -45,9 +47,7 @@ class NonSelectableMenu extends React.Component {
           <Menu.Item text="Default 1" key="1" className="TestFirstItem" />
           <Menu.Item text="Default 2" key="2" className="TestSecondItem" />
         </Menu>
-        <button type="button" id="non-selectable-menu-button" onClick={this.handleButtonClick} ref={this.setButtonNode}>
-          Default Menu
-        </button>
+        <Button id="non-selectable-menu-button" type="button" onClick={this.handleButtonClick} text="Default Menu" aria-haspopup icon={<IconCaretDown />} isReversed refCallback={this.setButtonNode} />
       </div>
     );
   }