+++ packages/terra-menu/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Fixed 
   * Fixed issue of navigation on menu header click.
+  
+* Added
+  * Allow screen readers to recognise as menu button
+  * Added context for menu item index, item with submenu and selected/unselected state
 
 ## 6.65.0 - (March 29, 2023)